creating a forecast SPA
The course is creating a simple two page forecast SPA using a service
to pass the city value between controllers.  In the course, he is using
$scope.$watch on the city.  Instead of that, I decided to add a
function to my service that updates the city.  I then inject the
service to my controller, attach the service to $scope, and use
ng-change to call the service update function.

I like this better than using $watch.  Since the service is responsible
for the current city, it should also bear the responsibility of
updating it as it changes.