simple weather app done
Separated out the pieces into different files.  This works fine for
such a small app, but I would change the structure if it were bigger.
I might go with the method of creating folders based on pieces.  So
there would be a folder for ‘dayForecast’ that holds the directive and
template for that day. And a folder for the index view and its
controller and so on.