put the weather stuff into a service
The weather fetching has now been moved to its own service, decoupling
it from the controller.  This would allow it to be easily used in
various locations.

I also moved the date and temp conversion into the controller for the
forecast directive since those are particular to a forecast (though a
date conversion really isn’t).  This way, instead of having to add so
many attributes to the directive, I just have to pass the day and the
date format.  A more robust version would allow for different temp
scales.  But for my purposes, this will do.